Jira issues fetching and react-routing changes
