+++ app/views/jiras/index.json.jbuilder
@@ -1,2 +1,4 @@
-# binding.pry
-json.array! @jiras, partial: "jiras/jira", as: :jira
+json.array! @issues do |issue|
+  json.key issue.key
+  json.summary issue.summary
+end
